Relocate Fruits/Flowers + cleanup strings
Relocated island/sister functionality to new FruitsFlowersEditor out of MiscPlayer to better follow existing layout. Added button to Map tab for the new form.
MiscPlayer fruit now sets passport only, FruitsFlowersEditor sets both. Other functionality remains as before.

Adjusted UI anchors and open positions + cleaned up recent lang_xx strings.
