+++ NHSE.WinForms/Resources/text/lang_de.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=Record Editor
 BuildingHelp=Building Help
 BulkSpawn=Bulk Spawn
+CampsiteEditor=Campingplatz Editor
 FieldItemEditor=Field Item Editor
 FlagEditor=Flag Editor
+FruitsFlowersEditor=Obst und Blumen-Editor
 ItemReceivedEditor=Received Item List Editor
 LandFlagEditor=Flag Editor
 MiscPlayerEditor=Misc Player Detail Editor
@@ -44,10 +46,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=Dump Villager
 Editor.B_EditAchievements=Edit Achievements
 Editor.B_EditBulletin=Edit Bulletin Board
+Editor.B_EditCampsite=Edit Campsite
 Editor.B_EditDesignsTailor=Edit Tailor Designs
 Editor.B_EditDIYTimer=Edit DIY Timer
 Editor.B_EditFieldGoods=Edit Field Goods
 Editor.B_EditFieldItems=Edit Field Items
+Editor.B_EditFruitFlower=Inselfrüchte und blumen bearbeiten
 Editor.B_EditFurniture=Edit Furniture
 Editor.B_EditHouses=Edit House
 Editor.B_EditLandFlags=Edit Flags
@@ -78,7 +82,6 @@ Editor.B_MoveOutAllVillagers=Box All Villagers
 Editor.B_RecycleBin=Edit Recycle Bin
 Editor.B_ReplaceVillager=Replace Villager
 Editor.B_SetPhraseOriginal=Original
-Editor.B_EditCampsite=Edit Campsite
 Editor.CHK_VillagerMovingOut=Moving Out
 Editor.L_AirportColor=Airport Color
 Editor.L_BankBells=Bank Bells:
@@ -115,6 +118,15 @@ Editor.Menu_Theme_Dark=Dark
 Editor.Tab_Map=Map
 Editor.Tab_Players=Players
 Editor.Tab_Villagers=Villagers
+CampsiteEditor.B_Cancel=Stornieren
+CampsiteEditor.B_Save=Speichern
+CampsiteEditor.L_Variant=Variante:
+CampsiteEditor.L_Species=Spezies:
+CampsiteEditor.L_Camper_ExternalName=Name des Besuchers:
+CampsiteEditor.L_Camper_InternalName=Dorfbewohner-Code:
+CampsiteEditor.L_Camper_Edit=Besucher bearbeiten:
+CampsiteEditor.L_CampsiteUnlockStatus=Der Campingplatz ist in diesem Speicherstand derzeit nicht freigeschaltet.
+CampsiteEditor.L_CampDate=Termin für den Campbesuch festlegen:
 FieldItemEditor.B_Cancel=Cancel
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=Dump Acre
@@ -227,6 +239,10 @@ FlagEditor.B_Dump=Dump
 FlagEditor.B_Load=Load
 FlagEditor.B_Save=Save
 FlagEditor.L_Count=Value:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=Obst (Lokal + Reisepass)
+FruitsFlowersEditor.L_ProfileSisterFruit=Obst (Schwester)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=Blume (Lokal)
+FruitsFlowersEditor.L_ProfileSisterFlower=Blume (Schwester)
 ItemReceivedEditor.B_AllArt=Give All Art
 ItemReceivedEditor.B_AllBugs=Give All Bugs
 ItemReceivedEditor.B_AllDive=Give All Dive
@@ -247,10 +263,7 @@ MiscPlayerEditor.B_Save=Save
 MiscPlayerEditor.CHK_CustomItem=Custom
 MiscPlayerEditor.CHK_ProfileMadeVillage=Made Village
 MiscPlayerEditor.L_Birthday=Birthday (M/D):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=Obst (Lokal)
-MiscPlayerEditor.L_ProfileSisterFruit=Obst (Schwester)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=Blume (Lokal)
-MiscPlayerEditor.L_ProfileSisterFlower=Blume (Schwester)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=Obst (Reisepass)
 MiscPlayerEditor.L_ProfileTimestamp=Timestamp
 MiscPlayerEditor.Tab_Misc=Misc
 MiscPlayerEditor.Tab_Profile=Profile
@@ -498,12 +511,3 @@ VillagerMemoryEditor.LBL_Name=Name:
 VillagerMemoryEditor.Tab_Flags=Flags
 VillagerMemoryEditor.Tab_Greet=Greet
 VillagerMemoryEditor.Tab_Misc=Misc
-CampsiteEditor.B_Cancel=Stornieren
-CampsiteEditor.B_Save=Speichern
-CampsiteEditor.L_Variant=Variante:
-CampsiteEditor.L_Species=Spezies:
-CampsiteEditor.L_Camper_ExternalName=Name des Besuchers:
-CampsiteEditor.L_Camper_InternalName=Dorfbewohner-Code:
-CampsiteEditor.L_Camper_Edit=Besucher bearbeiten:
-CampsiteEditor.L_CampsiteUnlockStatus=Der Campingplatz ist in diesem Speicherstand derzeit nicht freigeschaltet.
-CampsiteEditor.L_CampDate=Termin für den Campbesuch festlegen:
