+++ NHSE.WinForms/Resources/text/lang_es.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=Editor de Récords
 BuildingHelp=Ayuda de Construcción
 BulkSpawn=Aumentar Spawn
+CampsiteEditor=Editor de Camping
 FieldItemEditor=Editor de Objetos (Terreno)
 FlagEditor=Editor de Flags
+FruitsFlowersEditor=Editor de Frutas y Flores
 ItemReceivedEditor=Editor de Objetos Recibidos
 LandFlagEditor= Editor de Flags
 MuseumEditor=Editor de Flags
@@ -43,10 +45,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=Respaldar Ciudadano
 Editor.B_EditAchievements=Editar Logros
 Editor.B_EditBulletin=Editar Tablón de Anuncios
+Editor.B_EditCampsite=Edit Campsite
 Editor.B_EditDesignsTailor=Editar Diseños de Sastre
 Editor.B_EditDIYTimer=Edit DIY Timer
 Editor.B_EditFieldGoods=Editar Bienes (Terreno)
 Editor.B_EditFieldItems=Editar Objetos (Terreno)
+Editor.B_EditFruitFlower=Editar Frutas y Flores de la Isla
 Editor.B_EditFurniture=Editar Muebles
 Editor.B_EditHouses=Editar Casa
 Editor.B_EditLandFlags=Editar Flags
@@ -77,7 +81,6 @@ Editor.B_MoveOutAllVillagers=Box All Villagers
 Editor.B_RecycleBin=Editar Papelera de Reciclaje
 Editor.B_ReplaceVillager=Replace Villager
 Editor.B_SetPhraseOriginal=Original
-Editor.B_EditCampsite=Edit Campsite
 Editor.CHK_VillagerMovingOut=Moviebdo Afuera
 Editor.L_AirportColor=Color de Aeropuerto
 Editor.L_BankBells=Bayas (Banco):
@@ -114,6 +117,15 @@ Editor.Menu_Theme_Dark=Dark
 Editor.Tab_Map=Mapa
 Editor.Tab_Players=Jugadores
 Editor.Tab_Villagers=Ciudadanos
+CampsiteEditor.B_Cancel=Cancelar
+CampsiteEditor.B_Save=Guardar
+CampsiteEditor.L_Variant=Variante:
+CampsiteEditor.L_Species=Especies:
+CampsiteEditor.L_Camper_ExternalName=Nombre del visitante:
+CampsiteEditor.L_Camper_InternalName=Código de aldeano:
+CampsiteEditor.L_Camper_Edit=Editar Visitante:
+CampsiteEditor.L_CampsiteUnlockStatus=El camping no está desbloqueado actualmente en este guardado.
+CampsiteEditor.L_CampDate=Establecer fecha de visita al campamento:
 FieldItemEditor.B_Cancel=Cancelar
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=Guardar Acre
@@ -226,6 +238,10 @@ FlagEditor.B_Dump=Respaldar
 FlagEditor.B_Load=Cargar
 FlagEditor.B_Save=Guardar
 FlagEditor.L_Count=Valor:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=Fruta (Local + Pasaporte)
+FruitsFlowersEditor.L_ProfileSisterFruit=Fruta (Sister)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=Flor (Local)
+FruitsFlowersEditor.L_ProfileSisterFlower=Flor (Sister)
 ItemReceivedEditor.B_AllArt=Far Toda el Arte
 ItemReceivedEditor.B_AllBugs=Dar Todos los Bichos
 ItemReceivedEditor.B_AllDive=Give All Dive
@@ -246,10 +262,7 @@ MiscPlayerEditor.B_Save=Guardar
 MiscPlayerEditor.CHK_CustomItem=Custom
 MiscPlayerEditor.CHK_ProfileMadeVillage=Ciudadano Hecho
 MiscPlayerEditor.L_Birthday=Cumpleaños (M/D):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=Fruta (Local)
-MiscPlayerEditor.L_ProfileSisterFruit=Fruta (Sister)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=Flor (Local)
-MiscPlayerEditor.L_ProfileSisterFlower=Flor (Sister)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=Fruta (Pasaporte)
 MiscPlayerEditor.L_ProfileTimestamp=Timestamp
 MiscPlayerEditor.Tab_Misc=Misc.
 MiscPlayerEditor.Tab_Profile=Perfil
@@ -498,12 +511,3 @@ VillagerMemoryEditor.LBL_Name=Name:
 VillagerMemoryEditor.Tab_Flags=Flags
 VillagerMemoryEditor.Tab_Greet=Saludo
 VillagerMemoryEditor.Tab_Misc=Misc.
-CampsiteEditor.B_Cancel=Cancelar
-CampsiteEditor.B_Save=Guardar
-CampsiteEditor.L_Variant=Variante:
-CampsiteEditor.L_Species=Especies:
-CampsiteEditor.L_Camper_ExternalName=Nombre del visitante:
-CampsiteEditor.L_Camper_InternalName=Código de aldeano:
-CampsiteEditor.L_Camper_Edit=Editar Visitante:
-CampsiteEditor.L_CampsiteUnlockStatus=El camping no está desbloqueado actualmente en este guardado.
-CampsiteEditor.L_CampDate=Establecer fecha de visita al campamento:
