+++ NHSE.WinForms/Resources/text/lang_it.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=Modifica Record
 BuildingHelp=Aiuto Costruzione
 BulkSpawn=Genera in massa
+CampsiteEditor=Editore Campeggio
 FieldItemEditor=Modifica oggetti sul terreno
 FlagEditor=Modifica Flag
+FruitsFlowersEditor=Editore Frutta e Fiori
 ItemReceivedEditor=Modifica catalogo
 LandFlagEditor=Modifica Flag
 MiscPlayerEditor=Modifica dettagli giocatore
@@ -44,10 +46,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=Esporta Residente
 Editor.B_EditAchievements=Modifica Achievement
 Editor.B_EditBulletin=Modifica Albo
+Editor.B_EditCampsite=Modifica Campeggio
 Editor.B_EditDesignsTailor=Modifica Modelli Stilista
 Editor.B_EditDIYTimer=Edit DIY Timer
 Editor.B_EditFieldGoods=Modifica Beni Terreni
 Editor.B_EditFieldItems=Modifica Oggetti Sul Terreno
+Editor.B_EditFruitFlower=Modifica Frutta e Fiori dell'Isola
 Editor.B_EditFurniture=Modifica Oggetto Di Arredo
 Editor.B_EditHouses=Modifica Casa
 Editor.B_EditLandFlags=Modifica Flag Terreno
@@ -78,7 +82,6 @@ Editor.B_MoveOutAllVillagers=Box All Villagers
 Editor.B_RecycleBin=Modifica Cestino
 Editor.B_ReplaceVillager=Replace Villager
 Editor.B_SetPhraseOriginal=Original
-Editor.B_EditCampsite=Modifica Campeggio
 Editor.CHK_VillagerMovingOut=Trasloco Residente
 Editor.L_AirportColor=Colore Aeroporto
 Editor.L_BankBells=Stelline Banca:
@@ -115,6 +118,15 @@ Editor.Menu_Theme_Dark=Scuro
 Editor.Tab_Map=Mappa
 Editor.Tab_Players=Giocatori
 Editor.Tab_Villagers=Residenti
+CampsiteEditor.B_Cancel=Cancellare
+CampsiteEditor.B_Save=Salva
+CampsiteEditor.L_Variant=Variante:
+CampsiteEditor.L_Species=Specie:
+CampsiteEditor.L_Camper_ExternalName=Nome del visitatore:
+CampsiteEditor.L_Camper_InternalName=Codice del villico:
+CampsiteEditor.L_Camper_Edit=Modifica visitatore:
+CampsiteEditor.L_CampsiteUnlockStatus=Al momento, il campeggio non è sbloccato in questo salvataggio.
+CampsiteEditor.L_CampDate=Imposta la data della visita al campo:
 FieldItemEditor.B_Cancel=Indietro
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=Esporta Sezione
@@ -227,6 +239,10 @@ FlagEditor.B_Dump=Esporta
 FlagEditor.B_Load=Carica
 FlagEditor.B_Save=Salva
 FlagEditor.L_Count=Valore:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=Frutto (Locale + Passaporto)
+FruitsFlowersEditor.L_ProfileSisterFruit=Frutto (Sorella)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=Fiore (Locale)
+FruitsFlowersEditor.L_ProfileSisterFlower=Fiore (Sorella)
 ItemReceivedEditor.B_AllArt=Tutte le opere d'arte
 ItemReceivedEditor.B_AllBugs=Tutti gli insetti
 ItemReceivedEditor.B_AllDive=Give All Dive
@@ -247,10 +263,7 @@ MiscPlayerEditor.B_Save=Salva
 MiscPlayerEditor.CHK_CustomItem=Personale
 MiscPlayerEditor.CHK_ProfileMadeVillage=Isola Creata
 MiscPlayerEditor.L_Birthday=Compleanno (M/G):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=Frutto (Locale)
-MiscPlayerEditor.L_ProfileSisterFruit=Frutto (Sorella)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=Fiore (Locale)
-MiscPlayerEditor.L_ProfileSisterFlower=Fiore (Sorella)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=Frutto (Passaporto)
 MiscPlayerEditor.L_ProfileTimestamp=Registrazione
 MiscPlayerEditor.Tab_Misc=Altro
 MiscPlayerEditor.Tab_Profile=Profilo
@@ -499,12 +512,3 @@ VillagerMemoryEditor.Tab_Flags=Flags
 VillagerMemoryEditor.Tab_Greet=Saluti
 VillagerMemoryEditor.Tab_Misc=Altro
 VillagmoderMemoryEditor.L_Greeting3=Saluto 3:
-CampsiteEditor.B_Cancel=Cancellare
-CampsiteEditor.B_Save=Salva
-CampsiteEditor.L_Variant=Variante:
-CampsiteEditor.L_Species=Specie:
-CampsiteEditor.L_Camper_ExternalName=Nome del visitatore:
-CampsiteEditor.L_Camper_InternalName=Codice del villico:
-CampsiteEditor.L_Camper_Edit=Modifica visitatore:
-CampsiteEditor.L_CampsiteUnlockStatus=Al momento, il campeggio non è sbloccato in questo salvataggio.
-CampsiteEditor.L_CampDate=Imposta la data della visita al campo:
