+++ NHSE.WinForms/Resources/text/lang_jp.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=Record Editor
 BuildingHelp=Building Help
 BulkSpawn=Bulk Spawn
+CampsiteEditor=キャンプ場エディター
 FieldItemEditor=Field Item Editor
 FlagEditor=Flag Editor
+FruitsFlowersEditor=フルーツと花エディター
 ItemReceivedEditor=Received Item List Editor
 LandFlagEditor=Flag Editor
 MiscPlayerEditor=Misc Player Detail Editor
@@ -44,10 +46,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=Dump Villager
 Editor.B_EditAchievements=Edit Achievements
 Editor.B_EditBulletin=Edit Bulletin Board
+Editor.B_EditCampsite=キャンプ場を編集
 Editor.B_EditDesignsTailor=Edit Tailor Designs
 Editor.B_EditDIYTimer=Edit DIY Timer
 Editor.B_EditFieldGoods=Edit Field Goods
 Editor.B_EditFieldItems=Edit Field Items
+Editor.B_EditFruitFlower=島のフルーツと花を編集
 Editor.B_EditFurniture=Edit Furniture
 Editor.B_EditHouses=Edit House
 Editor.B_EditLandFlags=Edit Flags
@@ -78,7 +82,6 @@ Editor.B_MoveOutAllVillagers=Box All Villagers
 Editor.B_RecycleBin=Edit Recycle Bin
 Editor.B_ReplaceVillager=Replace Villager
 Editor.B_SetPhraseOriginal=Original
-Editor.B_EditCampsite=キャンプ場を編集
 Editor.CHK_VillagerMovingOut=Moving Out
 Editor.L_AirportColor=Airport Color
 Editor.L_BankBells=Bank Bells:
@@ -115,6 +118,15 @@ Editor.Menu_Theme_Dark=ダークモード
 Editor.Tab_Map=Map
 Editor.Tab_Players=Players
 Editor.Tab_Villagers=Villagers
+CampsiteEditor.B_Cancel=キャンセル
+CampsiteEditor.B_Save=保存
+CampsiteEditor.L_Variant=変異体：
+CampsiteEditor.L_Species=種：
+CampsiteEditor.L_Camper_ExternalName=村人の名前:
+CampsiteEditor.L_Camper_InternalName=村人コード:
+CampsiteEditor.L_Camper_Edit=訪問者を編集:
+CampsiteEditor.L_CampsiteUnlockStatus=このセーブではキャンプ場は現在ロック解除されていません。
+CampsiteEditor.L_CampDate=キャンプ訪問日を設定:
 FieldItemEditor.B_Cancel=Cancel
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=Dump Acre
@@ -227,6 +239,10 @@ FlagEditor.B_Dump=Dump
 FlagEditor.B_Load=Load
 FlagEditor.B_Save=Save
 FlagEditor.L_Count=Value:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=果物（地元+パスポート）
+FruitsFlowersEditor.L_ProfileSisterFruit=果物 (姉妹)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=花 (地元)
+FruitsFlowersEditor.L_ProfileSisterFlower=花 (姉妹)
 ItemReceivedEditor.B_AllArt=Give All Art
 ItemReceivedEditor.B_AllBugs=Give All Bugs
 ItemReceivedEditor.B_AllDive=Give All Dive
@@ -247,10 +263,7 @@ MiscPlayerEditor.B_Save=Save
 MiscPlayerEditor.CHK_CustomItem=Custom
 MiscPlayerEditor.CHK_ProfileMadeVillage=Made Village
 MiscPlayerEditor.L_Birthday=Birthday (M/D):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=果物（地元）
-MiscPlayerEditor.L_ProfileSisterFruit=果物 (姉妹)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=花 (地元)
-MiscPlayerEditor.L_ProfileSisterFlower=花 (姉妹)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=果物（パスポート）
 MiscPlayerEditor.L_ProfileTimestamp=Timestamp
 MiscPlayerEditor.Tab_Misc=Misc
 MiscPlayerEditor.Tab_Profile=Profile
@@ -498,12 +511,3 @@ VillagerMemoryEditor.LBL_Name=Name:
 VillagerMemoryEditor.Tab_Flags=Flags
 VillagerMemoryEditor.Tab_Greet=Greet
 VillagerMemoryEditor.Tab_Misc=Misc
-CampsiteEditor.B_Cancel=キャンセル
-CampsiteEditor.B_Save=保存
-CampsiteEditor.L_Variant=変異体：
-CampsiteEditor.L_Species=種：
-CampsiteEditor.L_Camper_ExternalName=村人の名前:
-CampsiteEditor.L_Camper_InternalName=村人コード:
-CampsiteEditor.L_Camper_Edit=訪問者を編集:
-CampsiteEditor.L_CampsiteUnlockStatus=このセーブではキャンプ場は現在ロック解除されていません。
-CampsiteEditor.L_CampDate=キャンプ訪問日を設定:
