+++ NHSE.WinForms/Resources/text/lang_ko.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=업적 편집기
 BuildingHelp=건물 도움말
 BulkSpawn=대량 생성
+CampsiteEditor=캠프장 편집기
 FieldItemEditor=필드 아이템 편집기
 FlagEditor=플래그 편집기
+FruitsFlowersEditor=과일 및 꽃 편집기
 ItemReceivedEditor=수령 아이템 목록 편집기
 LandFlagEditor=플래그 편집기
 MiscPlayerEditor=기타 플레이어 상세 편집기
@@ -44,10 +46,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=주민 덤프
 Editor.B_EditAchievements=업적 편집
 Editor.B_EditBulletin=게시판 편집
+Editor.B_EditCampsite=캠핑장 편집
 Editor.B_EditDesignsTailor=재봉소 디자인 편집
 Editor.B_EditDIYTimer=DIY 타이머 편집
 Editor.B_EditFieldGoods=필드 물품 편집
 Editor.B_EditFieldItems=필드 아이템 편집
+Editor.B_EditFruitFlower=섬 과일 및 꽃 편집
 Editor.B_EditFurniture=가구 편집
 Editor.B_EditHouses=집 편집
 Editor.B_EditLandFlags=플래그 편집
@@ -78,7 +82,6 @@ Editor.B_MoveOutAllVillagers=모든 주민 짐싸기
 Editor.B_RecycleBin=재활용함 편집
 Editor.B_ReplaceVillager=주민 교체
 Editor.B_SetPhraseOriginal=원본
-Editor.B_EditCampsite=캠핑장 편집
 Editor.CHK_VillagerMovingOut=이사 준비
 Editor.L_AirportColor=공항 색상
 Editor.L_BankBells=은행 벨:
@@ -115,6 +118,15 @@ Editor.Menu_Theme_Dark=다크 모드
 Editor.Tab_Map=지도
 Editor.Tab_Players=플레이어
 Editor.Tab_Villagers=주민
+CampsiteEditor.B_Cancel=취소
+CampsiteEditor.B_Save=구하다
+CampsiteEditor.L_Variant=변종:
+CampsiteEditor.L_Species=종:
+CampsiteEditor.L_Camper_ExternalName=방문자 이름:
+CampsiteEditor.L_Camper_InternalName=마을 주민 코드:
+CampsiteEditor.L_Camper_Edit=방문자 정보 수정:
+CampsiteEditor.L_CampsiteUnlockStatus=현재 이 저장 파일에서는 캠프장이 잠금 해제되지 않았습니다.
+CampsiteEditor.L_CampDate=캠프 방문 날짜를 정하세요:
 FieldItemEditor.B_Cancel=취소
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=에이커 덤프
@@ -227,6 +239,10 @@ FlagEditor.B_Dump=덤프
 FlagEditor.B_Load=불러오기
 FlagEditor.B_Save=저장
 FlagEditor.L_Count=값:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=과일(현지+여권)
+FruitsFlowersEditor.L_ProfileSisterFruit=과일 (자매)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=꽃 (현지)
+FruitsFlowersEditor.L_ProfileSisterFlower=꽃 (자매)
 ItemReceivedEditor.B_AllArt=모든 미술품 지급
 ItemReceivedEditor.B_AllBugs=모든 벌레 지급
 ItemReceivedEditor.B_AllDive=모든 해산물 지급
@@ -247,10 +263,7 @@ MiscPlayerEditor.B_Save=저장
 MiscPlayerEditor.CHK_CustomItem=사용자 지정
 MiscPlayerEditor.CHK_ProfileMadeVillage=마을 생성
 MiscPlayerEditor.L_Birthday=생일 (월/일):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=과일(현지)
-MiscPlayerEditor.L_ProfileSisterFruit=과일 (자매)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=꽃 (현지)
-MiscPlayerEditor.L_ProfileSisterFlower=꽃 (자매)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=과일(여권)
 MiscPlayerEditor.L_ProfileTimestamp=타임스탬프
 MiscPlayerEditor.Tab_Misc=기타
 MiscPlayerEditor.Tab_Profile=프로필
@@ -498,12 +511,3 @@ VillagerMemoryEditor.LBL_Name=이름:
 VillagerMemoryEditor.Tab_Flags=플래그
 VillagerMemoryEditor.Tab_Greet=인사
 VillagerMemoryEditor.Tab_Misc=기타
-CampsiteEditor.B_Cancel=취소
-CampsiteEditor.B_Save=구하다
-CampsiteEditor.L_Variant=변종:
-CampsiteEditor.L_Species=종:
-CampsiteEditor.L_Camper_ExternalName=방문자 이름:
-CampsiteEditor.L_Camper_InternalName=마을 주민 코드:
-CampsiteEditor.L_Camper_Edit=방문자 정보 수정:
-CampsiteEditor.L_CampsiteUnlockStatus=현재 이 저장 파일에서는 캠프장이 잠금 해제되지 않았습니다.
-CampsiteEditor.L_CampDate=캠프 방문 날짜를 정하세요:
