+++ NHSE.WinForms/Resources/text/lang_zhs.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=记录编辑器
 BuildingHelp=建筑帮助
 BulkSpawn=Bulk Spawn
+CampsiteEditor=营地编辑器
 FieldItemEditor=地皮编辑器
 FlagEditor=属性编辑器
+FruitsFlowersEditor=水果和鲜花编辑器
 ItemReceivedEditor=收藏清单编辑器
 LandFlagEditor=参数编辑器
 MiscPlayerEditor=杂项编辑器
@@ -44,10 +46,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=导出岛民
 Editor.B_EditAchievements=编辑成就
 Editor.B_EditBulletin=编辑布告栏
+Editor.B_EditCampsite=编辑露营地
 Editor.B_EditDesignsTailor=编辑裁缝设计
 Editor.B_EditDIYTimer=编辑DIY时间
 Editor.B_EditFieldGoods=编辑地表物品
 Editor.B_EditFieldItems=编辑地皮
+Editor.B_EditFruitFlower=编辑岛屿水果和鲜花
 Editor.B_EditFurniture=编辑家具
 Editor.B_EditHouses=编辑房屋
 Editor.B_EditLandFlags=编辑属性
@@ -78,7 +82,6 @@ Editor.B_MoveOutAllVillagers=赶走所有岛民
 Editor.B_RecycleBin=编辑回收箱
 Editor.B_ReplaceVillager=替换村民
 Editor.B_SetPhraseOriginal=初始口头禅
-Editor.B_EditCampsite=编辑露营地
 Editor.CHK_VillagerMovingOut=赶走岛民
 Editor.L_AirportColor=机场颜色
 Editor.L_BankBells=银行存款:
@@ -115,6 +118,15 @@ Editor.Menu_Theme_Dark=深色模式
 Editor.Tab_Map=地图
 Editor.Tab_Players=玩家
 Editor.Tab_Villagers=岛民
+CampsiteEditor.B_Cancel=取消
+CampsiteEditor.B_Save=节省
+CampsiteEditor.L_Variant=变体:
+CampsiteEditor.L_Species=物种：
+CampsiteEditor.L_Camper_ExternalName=访客姓名：
+CampsiteEditor.L_Camper_InternalName=村民代码：
+CampsiteEditor.L_Camper_Edit=编辑访客：
+CampsiteEditor.L_CampsiteUnlockStatus=此存档中露营地尚未解锁。
+CampsiteEditor.L_CampDate=确定营地参观日期：
 FieldItemEditor.B_Cancel=取消
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=导出耕地
@@ -227,6 +239,10 @@ FlagEditor.B_Dump=导出
 FlagEditor.B_Load=导入
 FlagEditor.B_Save=保存
 FlagEditor.L_Count=数值:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=水果（当地+护照）
+FruitsFlowersEditor.L_ProfileSisterFruit=水果 (姐妹)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=花 (当地)
+FruitsFlowersEditor.L_ProfileSisterFlower=花 (姐妹)
 ItemReceivedEditor.B_AllArt=获得所有美术品
 ItemReceivedEditor.B_AllBugs=获得所有虫子
 ItemReceivedEditor.B_AllDive=获得所有潜水服
@@ -247,10 +263,7 @@ MiscPlayerEditor.B_Save=保存
 MiscPlayerEditor.CHK_CustomItem=自定义
 MiscPlayerEditor.CHK_ProfileMadeVillage=已建立村庄
 MiscPlayerEditor.L_Birthday=生日 (月/日):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=水果（当地）
-MiscPlayerEditor.L_ProfileSisterFruit=水果 (姐妹)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=花 (当地)
-MiscPlayerEditor.L_ProfileSisterFlower=花 (姐妹)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=水果（护照）
 MiscPlayerEditor.L_ProfileTimestamp=登岛日
 MiscPlayerEditor.Tab_Misc=杂项
 MiscPlayerEditor.Tab_Profile=个人信息
@@ -498,12 +511,3 @@ VillagerMemoryEditor.LBL_Name=玩家名称:
 VillagerMemoryEditor.Tab_Flags=参数
 VillagerMemoryEditor.Tab_Greet=打招呼
 VillagerMemoryEditor.Tab_Misc=杂项
-CampsiteEditor.B_Cancel=取消
-CampsiteEditor.B_Save=节省
-CampsiteEditor.L_Variant=变体:
-CampsiteEditor.L_Species=物种：
-CampsiteEditor.L_Camper_ExternalName=访客姓名：
-CampsiteEditor.L_Camper_InternalName=村民代码：
-CampsiteEditor.L_Camper_Edit=编辑访客：
-CampsiteEditor.L_CampsiteUnlockStatus=此存档中露营地尚未解锁。
-CampsiteEditor.L_CampDate=确定营地参观日期：
