+++ NHSE.WinForms/Resources/text/lang_zht.txt
@@ -1,8 +1,10 @@
 ﻿AchievementEditor=成就編輯器
 BuildingHelp=建築說明
 BulkSpawn=Bulk Spawn
+CampsiteEditor=營地編輯器
 FieldItemEditor=地皮編輯器
 FlagEditor=屬性編輯器
+FruitsFlowersEditor=水果和鮮花編輯器
 ItemReceivedEditor=目錄清單編輯器
 LandFlagEditor=參數編輯器
 MiscPlayerEditor=雜項編輯器
@@ -44,10 +46,12 @@ BulkSpawn.L_Y=Y:
 Editor.B_DumpVillager=導出島民
 Editor.B_EditAchievements=編輯成就
 Editor.B_EditBulletin=編輯佈告欄
+Editor.B_EditCampsite=編輯露營地
 Editor.B_EditDesignsTailor=編輯裁縫設計
 Editor.B_EditDIYTimer=Edit DIY Timer
 Editor.B_EditFieldGoods=編輯地表物品
 Editor.B_EditFieldItems=編輯地皮
+Editor.B_EditFruitFlower=編輯島嶼水果和鮮花
 Editor.B_EditFurniture=編輯家具
 Editor.B_EditHouses=編輯房屋
 Editor.B_EditLandFlags=編輯屬性
@@ -78,7 +82,6 @@ Editor.B_MoveOutAllVillagers=Box All Villagers
 Editor.B_RecycleBin=編輯回收箱
 Editor.B_ReplaceVillager=Replace Villager
 Editor.B_SetPhraseOriginal=Original
-Editor.B_EditCampsite=編輯露營地
 Editor.CHK_VillagerMovingOut=趕走島民
 Editor.L_AirportColor=機場顏色
 Editor.L_BankBells=銀行存款:
@@ -115,6 +118,15 @@ Editor.Menu_Theme_Dark=深色模式
 Editor.Tab_Map=地圖
 Editor.Tab_Players=玩家
 Editor.Tab_Villagers=島民
+CampsiteEditor.B_Cancel=取消
+CampsiteEditor.B_Save=節省
+CampsiteEditor.L_Variant=變體：
+CampsiteEditor.L_Species=物種：
+CampsiteEditor.L_Camper_ExternalName=訪客姓名：
+CampsiteEditor.L_Camper_InternalName=村民代碼：
+CampsiteEditor.L_Camper_Edit=編輯訪客
+CampsiteEditor.L_CampsiteUnlockStatus=此存檔中露營地尚未解鎖。
+CampsiteEditor.L_CampDate=確定營地參觀日期：
 FieldItemEditor.B_Cancel=取消
 FieldItemEditor.B_Down=↓
 FieldItemEditor.B_DumpAcre=導出耕地
@@ -227,6 +239,10 @@ FlagEditor.B_Dump=導出
 FlagEditor.B_Load=導入
 FlagEditor.B_Save=儲存
 FlagEditor.L_Count=數值:
+FruitsFlowersEditor.L_ProfileSpecialtyFruit=水果（当地+護照）
+FruitsFlowersEditor.L_ProfileSisterFruit=水果 (姊妹)
+FruitsFlowersEditor.L_ProfileSpecialtyFlower=花 (当地)
+FruitsFlowersEditor.L_ProfileSisterFlower=花 (姊妹)
 ItemReceivedEditor.B_AllArt=獲得所有美術品
 ItemReceivedEditor.B_AllBugs=獲得所有蟲子
 ItemReceivedEditor.B_AllDive=Give All Dive
@@ -247,10 +263,7 @@ MiscPlayerEditor.B_Save=儲存
 MiscPlayerEditor.CHK_CustomItem=自訂
 MiscPlayerEditor.CHK_ProfileMadeVillage=已建立村莊
 MiscPlayerEditor.L_Birthday=生日 (月/日):
-MiscPlayerEditor.L_ProfileSpecialtyFruit=水果（当地）
-MiscPlayerEditor.L_ProfileSisterFruit=水果 (姊妹)
-MiscPlayerEditor.L_ProfileSpecialtyFlower=花 (当地)
-MiscPlayerEditor.L_ProfileSisterFlower=花 (姊妹)
+MiscPlayerEditor.L_ProfileSpecialtyFruit=水果（護照）
 MiscPlayerEditor.L_ProfileTimestamp=Timestamp
 MiscPlayerEditor.Tab_Misc=雜項
 MiscPlayerEditor.Tab_Profile=個人資訊
@@ -498,12 +511,3 @@ VillagerMemoryEditor.LBL_Name=Name:
 VillagerMemoryEditor.Tab_Flags=參數
 VillagerMemoryEditor.Tab_Greet=打招呼
 VillagerMemoryEditor.Tab_Misc=雜項
-CampsiteEditor.B_Cancel=取消
-CampsiteEditor.B_Save=節省
-CampsiteEditor.L_Variant=變體：
-CampsiteEditor.L_Species=物種：
-CampsiteEditor.L_Camper_ExternalName=訪客姓名：
-CampsiteEditor.L_Camper_InternalName=村民代碼：
-CampsiteEditor.L_Camper_Edit=編輯訪客
-CampsiteEditor.L_CampsiteUnlockStatus=此存檔中露營地尚未解鎖。
-CampsiteEditor.L_CampDate=確定營地參觀日期：
